fix(booking): address outbox review findings (CR-001~005,012)
- CR-001: Extract publishEvent into OutboxEventPublisher @Service
  to avoid self-invocation @Transactional bypass
- CR-002: Switch to native query with FOR UPDATE SKIP LOCKED
- CR-004: Polling query now locks rows to prevent concurrent processing
- CR-005: Add acks=all and enable.idempotence to outbox producer
- CR-012: Replace TODO with comment explaining reconciliation bypass
