+++ booking-service/src/main/java/com/sportstix/booking/event/outbox/OutboxPollingPublisher.java
@@ -1,16 +1,14 @@
 package com.sportstix.booking.event.outbox;
 
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import net.javacrumbs.shedlock.spring.annotation.SchedulerLock;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.concurrent.TimeUnit;
 
 /**
  * Polls outbox_events table and publishes pending events to Kafka.
@@ -18,21 +16,13 @@ import java.util.concurrent.TimeUnit;
  */
 @Slf4j
 @Component
+@RequiredArgsConstructor
 public class OutboxPollingPublisher {
 
     private static final int BATCH_SIZE = 50;
-    private static final int MAX_RETRIES = 5;
-    private static final int SEND_TIMEOUT_SECONDS = 5;
 
     private final OutboxEventRepository outboxEventRepository;
-    private final KafkaTemplate<String, String> kafkaTemplate;
-
-    public OutboxPollingPublisher(
-            OutboxEventRepository outboxEventRepository,
-            @Qualifier("outboxKafkaTemplate") KafkaTemplate<String, String> kafkaTemplate) {
-        this.outboxEventRepository = outboxEventRepository;
-        this.kafkaTemplate = kafkaTemplate;
-    }
+    private final OutboxEventPublisher outboxEventPublisher;
 
     @Scheduled(fixedDelay = 1000)
     @SchedulerLock(name = "outboxPolling", lockAtMostFor = "30s", lockAtLeastFor = "500ms")
@@ -45,35 +35,7 @@ public class OutboxPollingPublisher {
         log.debug("Polling {} outbox events", events.size());
 
         for (OutboxEvent event : events) {
-            publishEvent(event);
-        }
-    }
-
-    @Transactional
-    protected void publishEvent(OutboxEvent event) {
-        try {
-            kafkaTemplate.send(event.getTopic(), event.getPartitionKey(), event.getPayload())
-                    .get(SEND_TIMEOUT_SECONDS, TimeUnit.SECONDS);
-
-            event.markPublished();
-            outboxEventRepository.save(event);
-            log.debug("Outbox event published: id={}, topic={}", event.getId(), event.getTopic());
-        } catch (Exception e) {
-            handlePublishFailure(event, e);
-        }
-    }
-
-    private void handlePublishFailure(OutboxEvent event, Exception e) {
-        if (event.getRetryCount() >= MAX_RETRIES) {
-            event.markFailed();
-            outboxEventRepository.save(event);
-            log.error("Outbox event permanently failed after {} retries: id={}, topic={}",
-                    MAX_RETRIES, event.getId(), event.getTopic(), e);
-        } else {
-            event.markRetrying();
-            outboxEventRepository.save(event);
-            log.warn("Outbox event publish failed (retry {}): id={}, topic={}",
-                    event.getRetryCount(), event.getId(), event.getTopic(), e);
+            outboxEventPublisher.publishEvent(event);
         }
     }
